server/requirements.txt
Flask==2.2.5
Pillow==9.5.0
numpy==1.25.0
eth-account==0.8.1
